#8 [REFACTOR] UTILS 폴더구조 리팩토링
