+++ src/app/(auth)/signup/_funnel/steps/4-LoginInfo.tsx
@@ -9,8 +9,7 @@ import {
 import { useSignupStore } from "@/src/stores";
 import { useCheckLoginId } from "@/src/hooks/queries";
 import { usernameSchema, passwordSchema } from "@/src/schemas/signupSchema";
-import { validateField, validateMatch } from "@/src/utils/validation";
-import { formatLoginId } from "@/src/utils/format";
+import { validateField, validateMatch, formatLoginId } from "@/src/utils";
 import { SIGNUP_STEPS, SIGNUP_MESSAGES } from "@/src/constants/signup";
 import type { SignupStepProps } from "@/src/types";
