+++ src/utils/signup/format.ts
